Rename GraphHelpers, remove from legacy imports
