+++ app/assets/javascripts/helpers/GraphHelpers.js
